Explicitly define mutable strings
Ruby 2.3 introduced a magic comment, `frozen_string_literal: true`, that
allows strings to be immutable by default. Ruby 3.4 is preparing to make
this the default behavior and throws deprecation warnings when mutating
strings.

There are many ways to make a mutable string, but I've chosen to use
`String.new` because it is the most explicit option and also doesn't
require creating multiple string objects